+++ lib/src/exceptions/exceptions.dart
@@ -1,4 +1,29 @@
 /// Exception class representing errors related to credential operations.
+///
+/// ```dart
+///   code   message
+///    101   Initialization failure
+///    102   Plugin exception
+///    103   Not implemented
+///
+///    201   Login cancelled
+///    202   No credentials found
+///    203   Mismatched credentials
+///    204   Login failed
+///    205   Temporarily blocked (due to too many canceled sign-in prompts)
+///
+///    301   Save Credentials cancelled
+///    302   Create Credentials failed
+///
+///    401   Encryption failed
+///    402   Decryption failed
+///
+///    501   Received an invalid google id token response
+///    502   Invalid request
+///    503   Google client is not initialized yet
+///    504   Credentials operation failed
+///    505   Google credential decode error
+/// ```
 class CredentialException implements Exception {
   /// A numeric code identifying the specific error.
   final int code;
@@ -16,27 +41,3 @@ class CredentialException implements Exception {
     required this.details,
   });
 }
-
-/// Table for error codes and corresponding messages in [CredentialException].
-/*
-  code  message
-  101   Initialization failure
-  102   Plugin exception
-  103   Not implemented
-
-  201   Login cancelled
-  202   No credentials found
-  203   Mismatched credentials
-  204   Login failed
-
-  301   Save Credentials cancelled
-  302   Create Credentials failed
-
-  401   Encryption failed
-  402   Decryption failed
-
-  501  Received an invalid google id token response
-  502  Invalid request
-  503  Google client is not initialized yet
-  504  Credentials operation failed
-*/